Work-around stamper crashes related to Bitcoin proxy disconnects
Had some issues with the Bitcoin RPC proxy disconnecting, leading to a
BrokenPipeError exception that isn't caught properly. This appears to be due to
the Bitcoin Core RPC timing out and python-bitcoinlib not recovering; creating
a fresh connection should work around this issue.

It'd probably be better if python-bitcoinlib automatically re-opened the
connection for you, and/or Bitcoin Core be changed to not timeout.